refactor: move build_component_instance_for_tests utility to integration utils
