+++ src/backend/tests/integration/utils.py
@@ -7,6 +7,7 @@ import requests
 from astrapy.admin import parse_api_endpoint
 from langflow.api.v1.schemas import InputValueRequest
 from langflow.custom import Component
+from langflow.custom.eval import eval_custom_component_code
 from langflow.field_typing import Embeddings
 from langflow.graph import Graph
 from langflow.processing.process import run_graph_internal
@@ -180,3 +181,10 @@ async def run_single_component(
         graph, flow_id, session_id=session_id, inputs=graph_run_inputs, outputs=[component_id]
     )
     return graph.get_vertex(component_id).built_object
+
+
+def build_component_instance_for_tests(version: str, file_name: str = "Prompt", **kwargs):
+    component = download_component_from_github("prompts", file_name, version)
+    cc_class = eval_custom_component_code(component._code)
+    cc_instance = cc_class(**kwargs)
+    return cc_instance()
